requirements.txt
faiss_cpu==1.7.4
faiss_gpu==1.7.2
Flask==2.3.2
sentence_transformers==2.2.2
